Remove extended-join and invite-notify
The remaining caps only generate new commands which can easily be
filtered out when sending to clients so will be in the first pass of
support.

extended-join is probably safe to pass through unaltered, just causing
extraneous parameters on JOIN commands, but maybe not.

invite-notify reuses the INVITE command where the invited user is not
self.