Rename -A and -Q to -y and -q
The other upper-case options are related to the listening side of
things, not the server side, so this is more consistent.

This is incompatible, but will fail loudly, and I expect these options
are more likely set in a configuration file, if they are set at all. I
also want to free up -A for setting a client CA, but assuming your away
message is not also an existing file path, that will continue to fail
loudly.